Implement server and grading scheme in java
This is the basis for our server where it routes you to the index.html file. By clicking the link there, it takes you to cpe390_introhtml where you can input answers. The server will be activated on submit quiz button where it routes to TestQuizServlet. Here it creates a QuizToGrade which uses the other question classes to grade the quiz.