Add html and answer functions to questions
Adds the proper HTML in the buildhtml functions for the different questions. Also adds the basic functionality for printing answers from the buildAnswers function to the answer file.  Enables the use of Audio and video as well with the proper HTML added with the writeHTML functions in those display elements. Lastly, updates the TestQuiz file to use every question and display element type when testing.