Create preferences object and rearrange answer/id functions.
Creates a preferences object that will hold the users preferences for different aspects of the HTML to follow. This includes a hashmap for finding specific predefined keys. Question ids now get their initial letter from the abstract getqIdType() function which is implemented in each question child. Answers for multiple choice questions are also passed as strings with commas for separation and *s to identify correct answers.